split functional element constructor and instance
The types can no longer be assigned to each other.
The FunctionalElementConstructor interface was renamed to just Functionalelement.
The FunctionalElement interface generics now default to any so that multiple totally different FunctionalElement definitions
can be assigned to an array of plain FunctionalElement objects.
The FunctionalElementInstace interface generic now defaults to an empty object. (Previously it had no default.)
The defineFunctionalElement generics now default to an empty object rather than never. Using never was causing issues when
trying to assign a FunctionalElement definition with no initialized events (or props) to an array of plain Functionalelement.