add namedListen directive
This will listen to any event type by name. Thus, more complex event descriptions
(needed for the listen directive) are not needed. However, without these, extra
type information for the callback cannot be provided. When the event name is a
native DOM event, however, the event type will be inferred.